+++ CHANGELOG.md
@@ -4,11 +4,14 @@
    - All improvements below due to Martin Popel
    - Added evaluation on multiple test sets (e.g., `-t wmt17,wmt18`).
      Works as long as they all have the same language pair.
+   - Added `sacrebleu --origlang` (both for evaluation on a subset and for `--echo`).
+   - Added `sacrebleu --detail` for breakdown by subsets of the test set.
    - Minor changes
      - Improved display of `sacrebleu -h`
      - Added `sacrebleu --list`
      - Code refactoring
-     - Documentation updates
+     - Documentation and tests updates
+     - Fixed a race condition bug (`os.makedirs(outdir, exist_ok=True)` instead of `if os.path.exists`)
 
 - 1.3.7 (2019-07-12)
    - Lazy loading of regexes cuts import time from ~1s to nearly nothing (thanks, @louismartin!)